Move ground truth model to resemble an identity model
